minor revision

+++ include/svd.h
@@ -1,13 +1,14 @@
 #ifndef SVD_H
 #define SVD_H
 
-#include <vector>
-#include <utility>
 #include <Eigen/Dense>
+#include <utility>
+#include <vector>
 
 #include "point.h"
 
 namespace svd {
-    std::pair<Eigen::JacobiSVD<Eigen::MatrixXd>, Eigen::MatrixXd> compute(std::vector<Point>& points);
+std::pair<Eigen::JacobiSVD<Eigen::MatrixXf>, Eigen::MatrixXf> compute(
+    std::vector<Point>& points);
 };
 #endif /* SVD_H */
